chore: pinned items context provider and custom hook
